requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0
pandas>=2.0
numpy<2.0
scikit-learn>=1.3
statsmodels>=0.14
joblib>=1.3
requests>=2.31
python-dotenv>=1.0
tensorflow>=2.14.0
keras>=3.0.0
prophet>=1.1
streamlit>=1.28
schedule>=1.2
google-cloud-storage>=2.18
google-cloud-bigquery>=3.25
google-cloud-aiplatform>=1.67
